+++ CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Embedding plot: added a "No coloring" option to the "Color by" dropdown to disable coloring.
 - Show ground truth and prediction annotation sources in run details panel
 - Image details view: Annotations are now grouped by annotation source.
-
+- Image details view: Annotation source visibility can be toggled.
 - Image details: select the annotation source when adding an annotation.
 
 ### Changed